Refactors the project structure.
